Error handling using toast

+++ src/App.js
@@ -3,6 +3,8 @@ import { useState } from 'react';
 import './App.css';
 import FileUploadIcon from '@mui/icons-material/FileUpload';
 import DownloadIcon from '@mui/icons-material/Download';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 const App = () => {
 
@@ -26,6 +28,16 @@ const App = () => {
         ctx.drawImage(img, 0, 0);
       }
       img.onerror = () => {
+        toast.error('Please upload an image', {
+          position: "top-right",
+          autoClose: 3000,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: "light",
+        });
         console.error("The provided file couldn't be loaded as an Image media");
       }
       img.src = event.target.result;
@@ -33,27 +45,46 @@ const App = () => {
     reader.readAsDataURL(e.target.files[0]);
   }
 
+  const isNumeric = (value) => {
+    return /^-?\d+$/.test(value);
+  }
+
   const handleResize = () => {
-    let cv = window.cv;
-    let src = cv.imread('inputCanvas');
-    let dst = new cv.Mat();
-    let size = new cv.Size(parseInt(height), parseInt(width));
-    cv.resize(src, dst, size, 0, 0, cv.INTER_AREA);
-    cv.imshow('outputCanvas', dst);
-    setCanDownload(true);
-    src.delete(); dst.delete(); // to avoid memory leak
+    if (isNumeric(height) && isNumeric(width) && parseInt(height) > 0 && parseInt(width) > 0){
+      let cv = window.cv;
+      let src = cv.imread('inputCanvas');
+      let dst = new cv.Mat();
+      let size = new cv.Size(parseInt(height), parseInt(width));
+      cv.resize(src, dst, size, 0, 0, cv.INTER_AREA);
+      cv.imshow('outputCanvas', dst);
+      setCanDownload(true);
+      src.delete(); dst.delete(); // to avoid memory leak
+    } else {
+      toast.error('Please enter valid dimensions', {
+        position: "top-right",
+        autoClose: 3000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        theme: "light",
+      });
+    }
+    
   }
 
   const handleDownload = () => {
     var link = document.createElement('a');
     var name = imageName.substring(0, imageName.lastIndexOf("."));
-    link.download = 'resized_' + name + ".png";
+    link.download = 'resized_' + name + ".png"; // using the same image name to download
     link.href = document.getElementById('outputCanvas').toDataURL("image/png");
     link.click();
   }
 
   return (
     <div className="app">
+      <ToastContainer />
       <h1 className='heading'>Resize Image</h1>
       <Button className='uploadBtn' variant="contained" component="label" endIcon={<FileUploadIcon />}>
         {img == null ? "Upload Image" : "Upload new image"}
